refactoring, new init way of app, new structure, new config
